Create SM_GeneSetOverlap
This R script calculates the hypergeometric distribution for one gene set of interest and all other MSigDB signatures. It is necessary to download the MSigDB files prior to starting.

I made this script because I wanted to answer the question "which gene sets have the most significant overlap with my 'Downregulated_By_CCG-1423' gene set. But this script can be used with any geneset.


###Usage###
To use this script, put your favorite geneset in a .csv file so that the name of your gene set is in row 1, column 1, and all of the genes
in the gene set are in row 1, column 2, 3, 4...n+1. Save it as a CSV. In the following code, change 'CCG1423.csv' to whatever you named
your file. Ensure that you have downloaded the MSigDB datasets, and that they are named appropriately. It may be necessary to change the MSigDB file names, or the file names in the code.

It is not the most beautiful code, but it does its job. And since the calculations are not computationally intensive, the performance does
not matter that much, since it only takes seconds to run.

In the future I will add functionality so that it automatically generates and outputs a scatterplot using ggplot2.